+++ src/config.hpp
@@ -24,6 +24,7 @@ struct Config
   std::string ipc_endpoint;
   std::string url;
   unsigned int idle_timeout_seconds = 0;
+  uint8_t format_max = 1;
 
   // Attributes from CRSH_ATTR_*
   std::optional<std::string> bearer_token;